Extract fund code exception into a concern
